enclave: Extract OnlineWorkerSnapshot in system contract.
